+++ test/fixtures/form-vitals-data.js
@@ -12,213 +12,146 @@
 
 export const formVitalsData = [
   {
-    url: 'https://business.adobe.com/request-consultation/experience-cloud.html',
-    formViews: 8100,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
-      'desktop:mac': 100,
-    },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 8100,
-    pageviewsByUserAgent: {
-      'desktop:linux': 5800,
-      'mobile:ios': 300,
-      'desktop:windows': 1100,
-      'mobile:android': 100,
-      'desktop:mac': 800,
-    },
-  },
-  {
-    url: 'https://business.adobe.com/resources/reports/high-tech-digital-trends.html',
-    formViews: 400,
-    formSubmits: 200,
-    formSubmitByUserAgent: {
-      'desktop:windows': 100,
-      'mobile:ios': 100,
-    },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 400,
-    pageviewsByUserAgent: {
-      'desktop:windows': 200,
-      'mobile:ios': 100,
-      'desktop:mac': 100,
-    },
-  },
-  {
-    url: 'https://business.adobe.com/products/workfront/login.html',
-    formViews: 1500,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
-      'desktop:mac': 100,
-    },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 1500,
-    pageviewsByUserAgent: {
-      'desktop:windows': 600,
-      'mobile:android': 100,
-      'desktop:mac': 800,
-    },
-  },
-  {
-    url: 'https://business.adobe.com/resources/demo/adobe-learning-manager-lms.html',
-    formViews: 3400,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
+    url: 'https://business.adobe.com/se/resources/main.html',
+    formsubmit: {},
+    formview: {},
+    formengagement: {
       'desktop:windows': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 3400,
-    pageviewsByUserAgent: {
-      'desktop:windows': 1100,
-      'mobile:android': 1700,
-      'mobile:ios': 200,
-      'desktop:chromeos': 100,
-      'desktop:mac': 300,
+    pageview: {
+      'desktop:windows': 7898,
     },
   },
   {
-    url: 'https://business.adobe.com/it/products/workfront/login.html',
-    formViews: 100,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
+    url: 'https://business.adobe.com/jp/products/magento/magento-commerce.html',
+    formsubmit: {},
+    formview: {
       'desktop:mac': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 100,
-    pageviewsByUserAgent: {
+    formengagement: {},
+    pageview: {
       'desktop:mac': 100,
+      'desktop:linux': 2200,
+      'desktop:windows': 5890,
     },
   },
   {
-    url: 'https://business.adobe.com/request-consultation/genstudio.html',
-    formViews: 900,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
-      'desktop:mac': 100,
+    url: 'https://business.adobe.com/',
+    formsubmit: {},
+    formview: {},
+    formengagement: {
+      'desktop:windows': 100,
     },
-    errors: 100,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 900,
-    pageviewsByUserAgent: {
-      'desktop:windows': 700,
-      'desktop:mac': 100,
-      'mobile:android': 100,
+    pageview: {
+      'desktop:mac': 6789,
+      'desktop:windows': 1900,
+      'mobile:ios': 100,
+      'mobile:android': 300,
     },
   },
   {
     url: 'https://business.adobe.com/jp/resources/guides/leading-generative-ai-deployment-for-marketing.html',
-    formViews: 300,
-    formSubmits: 200,
-    formSubmitByUserAgent: {
+    formsubmit: {
       'desktop:windows': 200,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 300,
-    pageviewsByUserAgent: {
-      'desktop:windows': 300,
+    formview: {},
+    formengagement: {
+      'desktop:windows': 200,
+    },
+    pageview: {
+      'desktop:windows': 8000,
+    },
+  },
+  {
+    url: 'https://business.adobe.com/uk/products/magento/magento-commerce/sem.html',
+    formsubmit: {},
+    formview: {},
+    formengagement: {
+      'desktop:windows': 100,
+    },
+    pageview: {
+      'desktop:windows': 7890,
     },
   },
   {
     url: 'https://business.adobe.com/jp/resources/digital-trends-apj-report.html',
-    formViews: 100,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
+    formsubmit: {
       'desktop:windows': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 100,
-    pageviewsByUserAgent: {
+    formview: {},
+    formengagement: {
       'desktop:windows': 100,
+    },
+    pageview: {
+      'desktop:windows': 9765,
+    },
+  },
+  {
+    url: 'https://business.adobe.com/request-consultation/experience-cloud.html',
+    formsubmit: {
+      'desktop:mac': 100,
+    },
+    formview: {},
+    formengagement: {
+      'desktop:mac': 100,
+    },
+    pageview: {
+      'desktop:mac': 9786,
+      'desktop:linux': 1600,
     },
   },
   {
     url: 'https://business.adobe.com/products/workfront/get-demo.html',
-    formViews: 400,
-    formSubmits: 200,
-    formSubmitByUserAgent: {
+    formsubmit: {
       'desktop:mac': 100,
-      'mobile:android': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 400,
-    pageviewsByUserAgent: {
-      'desktop:windows': 200,
+    formview: {},
+    formengagement: {
       'desktop:mac': 100,
-      'mobile:android': 100,
+    },
+    pageview: {
+      'desktop:mac': 8656,
+    },
+  },
+  {
+    url: 'https://business.adobe.com/resources/demo/adobe-learning-manager-lms.html',
+    formsubmit: {
+      'desktop:windows': 100,
+    },
+    formview: {},
+    formengagement: {
+      'desktop:windows': 100,
+    },
+    pageview: {
+      'mobile:ios': 100,
+      'mobile:android': 200,
+      'desktop:windows': 7868,
     },
   },
   {
     url: 'https://business.adobe.com/jp/resources/reports/retail-digital-trends.html',
-    formViews: 100,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
+    formsubmit: {
       'desktop:windows': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 100,
-    pageviewsByUserAgent: {
+    formview: {},
+    formengagement: {
       'desktop:windows': 100,
+    },
+    pageview: {
+      'desktop:windows': 8668,
     },
   },
   {
     url: 'https://business.adobe.com/resources/reports/forrester-tei-adobe-content-supply-chain.html',
-    formViews: 200,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
+    formsubmit: {
       'desktop:mac': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 200,
-    pageviewsByUserAgent: {
-      'desktop:mac': 200,
-    },
-  },
-  {
-    url: 'https://business.adobe.com/cn/request-consultation/experience-cloud.html',
-    formViews: 100,
-    formSubmits: 100,
-    formSubmitByUserAgent: {
+    formview: {},
+    formengagement: {
       'desktop:mac': 100,
     },
-    errors: 0,
-    formSubmitButtonClicks: 0,
-    isFormViewPresent: false,
-    isFormSubmitPresent: true,
-    pageviews: 100,
-    pageviewsByUserAgent: {
-      'desktop:mac': 100,
+    pageview: {
+      'desktop:mac': 8678,
     },
   },
 ];